update: store remote config url in config file

+++ cmd/root.go
@@ -7,6 +7,9 @@ import (
 	"fmt"
 	"os"
 	"path"
+
+	"github.com/fatih/color"
+	"github.com/rodaine/table"
 )
 
 func Execute() {
@@ -25,6 +28,7 @@ func init() {
 
 	rootCmd.AddCommand(cleanDependencies)
 	cleanDependencies.Flags().BoolVarP(&skipConfirmation, "yes", "y", false, "Do not ask for confirmation before removing the dependencies")
+	cleanDependencies.Flags().BoolVarP(&isDryRun, "dry-run", "", false, "Prints all the file that will be deleted")
 
 	rootCmd.AddCommand(updateLocalConfig)
 
@@ -33,7 +37,7 @@ func init() {
 
 }
 
-func cleanupDir(dir string) {
+func cleanupDir(dir string, keepTheFile bool, filesToRemove *[]string) {
 	dir, isValid := validatePath(dir, false, true)
 
 	if !isValid {
@@ -49,12 +53,17 @@ func cleanupDir(dir string) {
 			continue
 		}
 
-		os.RemoveAll(fileSystemPath)
-	}
+		if !keepTheFile {
+			os.RemoveAll(fileSystemPath)
+		}
 
+		*filesToRemove = append(*filesToRemove, fileSystemPath)
+
+	}
 }
 
-func cleanDir(dir string) {
+func cleanDir(dir string) map[string]string {
+	cleanedDirs := map[string]string{}
 	dirsToClean := findAllChildDirs(dir)
 
 	dirsToCleanCount := len(dirsToClean)
@@ -63,12 +72,50 @@ func cleanDir(dir string) {
 		printError("Nothing to clean...")
 	}
 
-	bar := GenerateProgressBar(dirsToCleanCount, "Cleaning Dirs")
+	message := "Cleaning Dirs..."
+
+	if isDryRun {
+		message = "Scanning dirs..."
+	}
+
+	var filesToRemove []string
+
+	bar := GenerateProgressBar(dirsToCleanCount, message)
 	for i := 0; i < dirsToCleanCount; i++ {
 		dir := dirsToClean[i]
-		cleanupDir(dir)
+
+		cleanupDir(dir, isDryRun, &filesToRemove)
+
 		bar.Add(i + 1)
 	}
+	bar.Close()
+	fmt.Println()
+
+	if isDryRun {
+
+		headerFmt := color.New(color.FgGreen, color.Underline).SprintfFunc()
+		columnFmt := color.New(color.FgYellow).SprintfFunc()
+
+		tbl := table.New("FilePath")
+		tbl.WithHeaderFormatter(headerFmt).WithFirstColumnFormatter(columnFmt)
+
+		for i := range filesToRemove {
+			tbl.AddRow(filesToRemove[i])
+		}
+
+		fmt.Println()
+		tbl.Print()
+		fmt.Println()
+
+		promptForConfirmation("Above directories will be removed.")
+
+		for i := range filesToRemove {
+			os.RemoveAll(filesToRemove[i])
+		}
+
+	}
+
+	return cleanedDirs
 }
 
 func updateLocalConfigurationWithRemoteConfig() {
